Add two explicit casts to solve ambiguous overloads.

+++ src/interface/queue_storage.cpp
@@ -691,7 +691,7 @@ bool CQueueStorage::Impl::SaveServer(const CServerItem& item)
 	if (ret)
 	{
 		sqlite3_int64 serverId = sqlite3_last_insert_rowid(db_);
-		Bind(insertFileQuery_, file_table_column_names::server, serverId);
+		Bind(insertFileQuery_, file_table_column_names::server, static_cast<int64_t>(serverId));
 
 		const std::vector<CQueueItem*>& children = item.GetChildren();
 		for (std::vector<CQueueItem*>::const_iterator it = children.begin() + item.GetRemovedAtFront(); it != children.end(); ++it)
@@ -729,7 +729,7 @@ bool CQueueStorage::Impl::SaveFile(wxLongLong server, const CFileItem& file)
 
 	Bind(insertFileQuery_, file_table_column_names::download, file.Download() ? 1 : 0);
 	if (file.GetSize() != -1)
-		Bind(insertFileQuery_, file_table_column_names::size, file.GetSize().GetValue());
+		Bind(insertFileQuery_, file_table_column_names::size, static_cast<int64_t>(file.GetSize().GetValue()));
 	else
 		BindNull(insertFileQuery_, file_table_column_names::size);
 	if (file.m_errorCount)
